# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/api-config:
    get:
      operationId: getApiConfig
      tags: [settings]
      summary: Get API config
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiConfig"
    put:
      operationId: updateApiConfig
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiConfigUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiConfig"
  /settings/test-connection/{service}:
    post:
      operationId: testConnection
      tags: [settings]
      parameters:
        - name: service
          in: path
          required: true
          schema:
            type: string
            enum: [bolna, brevo, meta]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestConnectionResult"
...
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: policyType
          in: query
          required: false
          schema: { $ref: "#/components/schemas/InsuranceType" }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/PolicyStatus" }
        - name: insurer
          in: query
          required: false
          schema: { type: string }
        - name: renewalFrom
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: renewalTo
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /policies/{id}:
    get:
      operationId: getPolicy
      tags: [policies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    patch:
      operationId: updatePolicy
      tags: [policies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
    delete:
      operationId: deletePolicy
      tags: [policies]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
...
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: inviteTeamMember
      tags: [team]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    delete:
      operationId: removeTeamMember
      tags: [team]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    InsuranceType:
      type: string
      enum: [LIFE, HEALTH, MOTOR, TERM, ULIP, ENDOWMENT, ACCIDENT, TRAVEL]
...
    PolicyStatus:
      type: string
      enum: [ACTIVE, LAPSED, SURRENDERED, MATURED, CLAIMED]
    PremiumFreq:
      type: string
      enum: [MONTHLY, QUARTERLY, HALF_YEARLY, YEARLY, SINGLE]
...
    ApiConfig:
      type: object
      required: [id, org_id, bolna_base_url, webhook_secret, context_api_bearer_token, context_api_url, meta_webhook_url, website_form_webhook_url]
      properties:
        id: { type: string }
        org_id: { type: string }
        bolna_api_key: { type: ["string", "null"] }
        bolna_base_url: { type: string }
        brevo_api_key: { type: ["string", "null"] }
        brevo_sender_name: { type: ["string", "null"] }
        meta_ads_access_token: { type: ["string", "null"] }
        meta_ads_account_id: { type: ["string", "null"] }
        webhook_secret: { type: string }
        context_api_bearer_token: { type: string }
        monthly_checkin_agent_id: { type: ["string", "null"] }
        sms_on_lead_created: { type: boolean }
        sms_on_call_scheduled: { type: boolean }
        email_renewal_reminders: { type: boolean }
        context_api_url: { type: string }
        meta_webhook_url: { type: string }
        website_form_webhook_url: { type: string }
        updated_at: { type: string, format: date-time }
    ApiConfigUpdate:
      type: object
      properties:
        bolna_api_key: { type: string }
        bolna_base_url: { type: string }
        brevo_api_key: { type: string }
        brevo_sender_name: { type: string }
        meta_ads_access_token: { type: string }
        meta_ads_account_id: { type: string }
        monthly_checkin_agent_id: { type: string }
        sms_on_lead_created: { type: boolean }
        sms_on_call_scheduled: { type: boolean }
        email_renewal_reminders: { type: boolean }
    TestConnectionResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
...
    Policy:
      type: object
      required: [id, org_id, lead_id, policy_type, premium_frequency, status, created_at, updated_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        lead_id: { type: string }
        lead_name: { type: ["string", "null"] }
        policy_number: { type: ["string", "null"] }
        insurer_name: { type: ["string", "null"] }
        policy_type: { $ref: "#/components/schemas/InsuranceType" }
        sum_assured: { type: ["integer", "null"] }
        annual_premium: { type: ["integer", "null"] }
        premium_frequency: { $ref: "#/components/schemas/PremiumFreq" }
        start_date: { type: ["string", "null"], format: date-time }
        end_date: { type: ["string", "null"], format: date-time }
        renewal_date: { type: ["string", "null"], format: date-time }
        nominee_name: { type: ["string", "null"] }
        nominee_relation: { type: ["string", "null"] }
        nominee_dob: { type: ["string", "null"], format: date-time }
        status: { $ref: "#/components/schemas/PolicyStatus" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/PolicyDocument" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PolicyDocument:
      type: object
      required: [name, url]
      properties:
        name: { type: string }
        url: { type: string }
    PolicyInput:
      type: object
      required: [lead_id, policy_type]
      properties:
        lead_id: { type: string }
        policy_number: { type: string }
        insurer_name: { type: string }
        policy_type: { $ref: "#/components/schemas/InsuranceType" }
        sum_assured: { type: integer }
        annual_premium: { type: integer }
        premium_frequency: { $ref: "#/components/schemas/PremiumFreq" }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        renewal_date: { type: string, format: date-time }
        nominee_name: { type: string }
        nominee_relation: { type: string }
        nominee_dob: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/PolicyStatus" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/PolicyDocument" }
    PolicyUpdate:
      type: object
      properties:
        policy_number: { type: ["string", "null"] }
        insurer_name: { type: ["string", "null"] }
        policy_type: { $ref: "#/components/schemas/InsuranceType" }
        sum_assured: { type: ["integer", "null"] }
        annual_premium: { type: ["integer", "null"] }
        premium_frequency: { $ref: "#/components/schemas/PremiumFreq" }
        start_date: { type: ["string", "null"], format: date-time }
        end_date: { type: ["string", "null"], format: date-time }
        renewal_date: { type: ["string", "null"], format: date-time }
        nominee_name: { type: ["string", "null"] }
        nominee_relation: { type: ["string", "null"] }
        nominee_dob: { type: ["string", "null"], format: date-time }
        status: { $ref: "#/components/schemas/PolicyStatus" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/PolicyDocument" }
...
    TeamMember:
      type: object
      required: [id, org_id, name, email, role, created_at]
      properties:
        id: { type: string }
        org_id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        created_at: { type: string, format: date-time }
    TeamMemberInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        role: { type: string }